requirements.txt
PyQt5==5.7
pyserial==3.2.1
sip==4.18.1
